`linera-rpc`: replace `generate-format` with `insta`
