--- linera-rpc/generate-format.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-cd "`dirname $0`"/..
-cargo run --example generate-format > linera-rpc/tests/staged/formats.yaml
